updated license_grant & base price

+++ src/main/java/com/i3market/semanticengine/common/domain/entity/PricingModel.java
@@ -17,7 +17,7 @@ public class PricingModel implements Serializable {
 
     String pricingModelName;
 
-    BigDecimal basicPrice;
+    float basicPrice;
 
     String currency;
