+++ src/main/java/com/i3market/semanticengine/common/domain/response/PricingModelDto.java
@@ -25,7 +25,7 @@ public class PricingModelDto {
 
     String pricingModelName;
 
-    BigDecimal basicPrice;
+    float basicPrice;
 
     String currency;
